Drop explicit "rails" dependency in favour of slimmer "railties"
The mimemagic incident [1] has shown that it's always good to require as
few dependencies as possible. A Rails Engine only needs the `railties`
gem, plus whatever else it does internally, so instead of pulling in the
whole of Rails for our gem we can just pick and choose the parts we
want.

[1]: https://twitter.com/nateberkopec/status/1374722404228853762