Always destroy old token after successfully calling #refresh!
Once a refresh token has been used, it cannot be re-used, so expired
tokens will remain in an invalid state forever until the situation is
rectified elsewhere.

Therefore, we should always delete the old token immediately, regardless
of whether the replacement was persisted, because there was a side
effect outside the transaction (the Hub will no longer accept that
refresh token) and we must update our system to match.