Don't return access tokens which expire in <30s, and allow setting this
I encounted an issue where we fetched an access token which was due to
expire in one second. Since my usage is catching all HTTP errors, I
would need to detect the 403-specific error and do something
differently.

Instead, it would be better to specify a "lease time" when getting the
token - I know approximately how long I would need the token for, so I
will request a token which has enough "lease" left on it to fulfil my
needs.

This change will probably lead to a small increase in the number of
tokens in the system, because we'll be proactively refreshing tokens
before their official expiration. However, this tradeoff means that it
is less likely to encounter the issue of an expired token mid-process.