Modify AccessToken.usable scope to consider validity periods
The previous code would fetch a token even if it would expire within 30
seconds (by default), and then attempt to refresh it. If the token was
received via client_credentials flow, it has no refresh token, which
throws an error.

In addition, it is still possible (but unlikely) to receive a token
which needs refreshing but cannot be refreshed. So the interface to
`#refresh!` now returns if `refresh_token` is nil - the handler would
then attempt to get a fresh token via client credentials instead.