Set `allow_other_host: true` when redirecting to Hub
This is required in Rails 7 if using the
`ActionController::Base.raise_on_open_redirects` configuration flag
(which is the default). Otherwise, Rails will throw an
`ActionController::Redirecting::UnsafeRedirectError` for us.

I think these are the only two places where this client redirects _away_
from the parent application to the Hub - the other redirects are
presumably internal.